requirements.txt
flask==2.0.1
transformers==4.30.2
pandas==1.3.3
requests==2.31.0
python-dotenv==0.19.0 